이미지를 다운로드 하는 역할을 StoreItems가 하도록 수정
