+++ StoreApp/StoreApp/Models/StoreItems.swift
@@ -72,12 +72,41 @@ extension JSONDataFetcher {
             completion(items)
         }
     }
+    
+    func downloadImage(with url: URL) {
+        URLSession.shared.downloadTask(with: url) { (fileURL, response, error) in
+            if let error = error {
+                print(error.localizedDescription)
+                return
+            }
+            
+            guard let fileURL = fileURL,
+                let response = response as? HTTPURLResponse,
+                (200...299) ~= response.statusCode,
+                let cacheURL = URL.cachesDirectory() else { return }
+            
+            let savedURL = cacheURL.appendingPathComponent(url.lastPathComponent)
+            try? FileManager.default.moveItem(at: fileURL, to: savedURL)
+            
+            let userInfo = [UserInfoKey.imageName: url.lastPathComponent]
+            NotificationCenter.default.post(name: .imageDidDownload,
+                                            object: nil,
+                                            userInfo: userInfo)
+            
+            }.resume()
+    }
 }
 
 extension Array where Element == StoreItem {
     func downloadImages() {
         for item in self {
-            item.downloadImage()
+            let image = item.image
+            guard let imageURL = URL(string: image),
+                let cacheURL = URL.cachesDirectory() else { return }
+            let imagePath = cacheURL.appendingPathComponent(image.lastPathComponent())
+            guard FileManager.default.fileExists(atPath: imagePath.path) == false else { return }
+            let jsonFetcher = JSONDataFetcher()
+            jsonFetcher.downloadImage(with: imageURL)
         }
     }
 }
